requirements.txt
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
awscli==1.42.0
axios==0.4.0
azure-ai-agents==1.0.1
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b12
azure-core==1.34.0
azure-cosmos==4.9.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
bcrypt==4.3.0
boto3==1.40.0
botocore==1.40.0
Brotli==1.1.0
CacheControl==0.14.3
cachetools==5.5.2
cadquery==2.5.2
casadi==3.7.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cssselect2==0.8.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
docutils==0.19
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
ezdxf==1.4.2
fastapi==0.115.14
firebase-admin==6.9.0
fonttools==4.58.4
gmsh==4.14.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.109.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.23.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
kiwisolver==1.4.8
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
meshio==5.3.5
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.1.1
multimethod==1.12
mypy_extensions==1.1.0
nlopt==2.9.1
numpy==2.2.6
openai==1.93.0
packaging==25.0
path==17.1.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydyf==0.11.0
pygltflib==1.16.4
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.4
rich==14.0.0
rsa==4.7.2
s3transfer==0.13.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
stripe==12.5.1
tenacity==8.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
typish==1.9.3
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
zopfli==0.2.3.post1
resend>=2,<3
